Delete unused MarginCalculator library
Update tests to test the margin calculation logic that we are actulaly using, in the MarginEngine.

Note that when margin calculation code was moved into the MarginEngine, the order of some math operations changed very slightly to improve accuracy). THis commit temporarily reverts to the old math operation order, to aid getting the tests to pass in a new context.

The new test code changes start/end timestamps, and manipulates the block time, to get deterministic results that are equivalent to the old tests.

TODO:
 - reinstate slightly more accurate lambdaWad calculation and update test results so that tests pass, oncme we know everythign else is equivalent